[sdlf-datalakeLibrary] remove legacy octagon support
remove octagon configuration/metadata and get configuration exclusively from SSM Parameter Store
this also allows merging configuration/ into sdlf/ (previously octagon/)

simplify datalakeLibrary - OctagonClient and PipelineExecutionHistoryAPI were needlessly separate

update SSM parameter names to fetch the newest format
