+++ sdlf-datalakeLibrary/python/datalake_library/interfaces/dynamo_interface.py
@@ -1,4 +1,5 @@
 import datetime as dt
+import logging
 import os
 from types import TracebackType
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Type
@@ -17,6 +18,7 @@ if TYPE_CHECKING:
         WriteRequestTypeDef,
     )
 
+logger = logging.getLogger(__name__)
 
 class DynamoInterface:
     def __init__(self, configuration, log_level=None, dynamodb_client=None):
@@ -440,3 +442,19 @@ class _TableBatchWriter:
             self._flush()
 
         return None
+
+def clean_table(dynamodb, table_name, pk_name, sk_name=""):
+    logger.debug(f"Clean dynamodb table {table_name}, PK: {pk_name}, SK: {sk_name}")
+    table = dynamodb.Table(table_name)
+    while True:
+        result = table.scan()
+        if result["Count"] == 0:
+            logger.debug(f"Clean dynamodb table {table_name}... DONE")
+            return
+
+        with table.batch_writer() as batch:
+            for item in result["Items"]:
+                if not sk_name:
+                    batch.delete_item(Key={pk_name: item[pk_name]})
+                else:
+                    batch.delete_item(Key={pk_name: item[pk_name], sk_name: item[sk_name]})
