--- sdlf-datalakeLibrary/python/datalake_library/octagon/__version__.py
@@ -1,4 +1,0 @@
-# Octagon Python API
-
-__title__ = "Octagon"
-__version__ = "1.0"
