+++ sdlf-datalakeLibrary/python/datalake_library/sdlf/__init__.py
@@ -1,11 +1,9 @@
 import logging
 
 from .__version__ import __title__, __version__  # noqa: F401;
-from .artifact import Artifact  # noqa: F401;
-from .client import OctagonClient  # noqa: F401;
-from .event import EventReasonEnum  # noqa: F401;
+from .client import SdlfClient  # noqa: F401;
 
-name = "octagon"
+name = "sdlf"
 
 # Suppress boto3 logging
 logging.getLogger("boto3").setLevel(logging.CRITICAL)
