+++ sdlf-datalakeLibrary/python/datalake_library/sdlf/__version__.py
@@ -1,0 +1,2 @@
+__title__ = "SDLF"
+__version__ = "2.7.0a1"
