+++ sdlf-datalakeLibrary/python/datalake_library/sdlf/peh.py
@@ -1,9 +1,13 @@
 import datetime
 import logging
+import os
 import uuid
 from decimal import Decimal
 
+import boto3
 from boto3.dynamodb.types import TypeSerializer
+
+from .config import DynamoConfiguration
 
 from ..commons import deserialize_dynamodb_item, serialize_dynamodb_item
 from .utils import (
@@ -28,25 +32,74 @@ PEH_STATUS_CANCELED = "CANCELED"
 class PipelineExecutionHistoryAPI:
     pipelines = dict()  # Pipelines cache across all instances
 
-    def __init__(self, client):
-        self.logger = logging.getLogger(__name__)
-        self.client = client
-        self.dynamodb = client.dynamodb
-        self.pipelines_table = client.config.get_pipelines_table()
-        self.peh_table = client.config.get_peh_table()
-        self.peh_ttl = client.config.get_peh_ttl()
+    def __init__(
+            self,
+            run_in_context,
+            region: str = "us-east-1",
+            profile: str = "default",
+            instance: str = "dev",
+            sns_topic: str = None,
+        ):
+        self.dynamodb = boto3.client("dynamodb")
+        dynamo_config = DynamoConfiguration()
+        self.peh_table = dynamo_config.peh_table
+        self.peh_ttl = 120 # TODO property of dynamo_config?
 
-    def start_pipeline_execution(self, pipeline_name, dataset_name=None, dataset_date=None, comment=None):
-        self.logger.debug("peh start_pipeline_execution() called")
+        self._logger = logging.getLogger(__name__)
+
+        self.region = region
+        self.profile = profile
+        self.run_in_fargate = run_in_context == "FARGATE"
+        self.run_in_lambda = run_in_context == "LAMBDA"
+        self.sns_topic = sns_topic
+
+        # No current pipeline execution
+        self.pipeline_name = None
+        self.pipeline_execution_id = None
+
+        if self.run_in_fargate:
+            if "AWS_ACCESS_KEY" in os.environ and "AWS_SECRET_ACCESS_KEY" in os.environ:
+                aws_access_key = os.getenv("AWS_ACCESS_KEY")
+                aws_secret_access_key = os.getenv("AWS_SECRET_ACCESS_KEY")
+                boto3.setup_default_session(
+                    profile_name=self.profile,
+                    region_name=self.region,
+                    aws_access_key_id=aws_access_key,
+                    aws_secret_access_key=aws_secret_access_key,
+                )
+            else:
+                msg = "Environment variables AWS_ACCESS_KEY and AWS_SECRET_ACCESS_KEY are not set"
+                self._logger.error(msg)
+                raise ValueError(msg)
+        elif self.run_in_lambda:
+            pass
+        else:
+            boto3.setup_default_session(profile_name=self.profile, region_name=self.region)
+
+        sts_endpoint_url = "https://sts." + os.getenv("AWS_REGION") + ".amazonaws.com"
+        self.account_id = boto3.client("sts", endpoint_url=sts_endpoint_url).get_caller_identity().get("Account")
+
+
+    def start_pipeline_execution(self, pipeline_name: str, dataset_name: str = None, dataset_date: str = None, comment: str = None) -> str:
+        """Creates a record for Pipeline Execution History
+
+        Arguments:
+            pipeline_name {str} -- Name of pipeline (needs to be active)
+
+        Keyword Arguments:
+            dataset_date {str} -- Optional. ISO 8601 date of dataset
+            comment {str} -- Optional. Comment
+
+        Returns:
+            str -- Unique reference to a Pipeline Execution History record (uuid4)
+        """
+
+        self._logger.debug("peh start_pipeline_execution() called")
 
         throw_none_or_empty(pipeline_name, "Pipeline name is not specified")
 
         if dataset_date:
             validate_date(dataset_date)
-
-        if not self.check_pipeline(pipeline_name):
-            self.logger.error(f"Pipeline doesn't exist or inactive : {pipeline_name}")
-            return None
 
         peh_id = str(uuid.uuid4())
         current_time = datetime.datetime.now(datetime.UTC)
@@ -88,15 +141,24 @@ class PipelineExecutionHistoryAPI:
         serializer = TypeSerializer()
         self.dynamodb.put_item(TableName=self.peh_table, Item=serialize_dynamodb_item(item, serializer))
 
-        self.client.set_pipeline_execution(peh_id, pipeline_name)
+        self.set_pipeline_execution(peh_id, pipeline_name)
 
         return peh_id
 
-    def update_pipeline_execution(self, status, component=None, issue_comment=None):
-        self.logger.debug("peh create_execution() called")
+    def update_pipeline_execution(self, status: str, component: str = None, issue_comment: str = None):
+        """Update status of Pipeline Execution History record
 
-        throw_if_false(self.client.is_pipeline_set(), "Pipeline execution is not yet assigned")
-        peh_id = self.client.pipeline_execution_id
+        Arguments:
+            status {str} -- New status of Pipeline Execution
+            component {str} -- Optional. Component of Pipeline Execution
+
+        Returns:
+            bool -- True if successful
+        """
+        self._logger.debug("peh create_execution() called")
+
+        throw_if_false(self.is_pipeline_set(), "Pipeline execution is not yet assigned")
+        peh_id = self.pipeline_execution_id
 
         peh_rec = self.get_peh_record(peh_id)
         if peh_rec:
@@ -190,7 +252,7 @@ class PipelineExecutionHistoryAPI:
                 expr_values[":C"] = issue_comment
                 update_expr += ", #C = :C"
 
-        # self.logger.debug(f"Update: {update_expr} \nNames: {expr_names} \nValues{expr_values}")
+        # self._logger.debug(f"Update: {update_expr} \nNames: {expr_names} \nValues{expr_values}")
 
         serializer = TypeSerializer()
         self.dynamodb.update_item(
@@ -202,88 +264,27 @@ class PipelineExecutionHistoryAPI:
             ReturnValues="UPDATED_NEW",
         )
 
-        # Add pipeline update for COMPLETED Executions
-        if status == PEH_STATUS_COMPLETED:
-            self.logger.debug(f"Pipeline: {self.client.pipeline_name}")
-            item = deserialize_dynamodb_item(
-                self.dynamodb.get_item(
-                    TableName=self.pipelines_table,
-                    Key={"name": {"S": self.client.pipeline_name}},
-                    ConsistentRead=True,
-                    ProjectionExpression="#n, version",
-                    ExpressionAttributeNames={"#n": "name"},
-                )["Item"]
-            )
-            pipeline_version = item["version"]
-
-            expr_names = {
-                "#V": "version",
-                "#U": "last_updated_timestamp",
-                "#P": "last_execution_id",
-                "#D": "last_execution_date",
-                "#E": "last_execution_timestamp",
-                "#S": "last_execution_status",
-                "#X": "last_execution_duration_in_seconds",
-            }
-
-            expr_values = {
-                ":V": pipeline_version,
-                ":INC": 1,
-                ":S": status,
-                ":P": self.client.pipeline_execution_id,
-                ":D": local_date_iso,
-                ":E": utc_time_iso,
-                ":U": utc_time_iso,
-                ":X": Decimal(str(duration_sec)),
-            }
-            update_expr = "SET #P = :P, #V = :V + :INC, #S = :S, #D = :D, #X = :X, #E = :E, #U = :U"
-
-            serializer = TypeSerializer()
-            self.dynamodb.update_item(
-                TableName=self.pipelines_table,
-                Key={"name": {"S": self.client.pipeline_name}},
-                UpdateExpression=update_expr,
-                ExpressionAttributeNames=expr_names,
-                ExpressionAttributeValues=serialize_dynamodb_item(expr_values, serializer),
-                ReturnValues="UPDATED_NEW",
-            )
-
         return True
 
     def get_peh_record(self, peh_id):
-        # self.logger.debug(f"check_peh_active(): {peh_id}")
+        # self._logger.debug(f"check_peh_active(): {peh_id}")
         result = self.dynamodb.get_item(TableName=self.peh_table, Key={"id": {"S": peh_id}}, ConsistentRead=True)
-        # self.logger.debug(f"check_peh_active(): {result}")
+        # self._logger.debug(f"check_peh_active(): {result}")
 
         if "Item" in result:
             return deserialize_dynamodb_item(result["Item"])
         else:
             return None
 
-    # Check if pipeline exists and active
-    def check_pipeline(self, pipeline_name: str) -> bool:
-        self.logger.debug(f"check_pipeline: {pipeline_name}")
-        if pipeline_name not in self.pipelines.keys():  # Pipeline not found in cache
-            # Go to DDB and add new pipeline to cache
-            self.logger.debug(f"check_pipeline - get from DDB: {pipeline_name}")
-            result = self.dynamodb.get_item(
-                TableName=self.pipelines_table,
-                Key={"name": {"S": pipeline_name}},
-                ConsistentRead=True,
-                ProjectionExpression="#n, #s",
-                ExpressionAttributeNames={"#n": "name", "#s": "status"},
-            )
-            self.logger.debug("result:" + str(result))
-            if "Item" in result:  # Pipeline found, check status
-                status = deserialize_dynamodb_item(result["Item"])["status"]
-                self.pipelines[pipeline_name] = status == PIPELINE_STATUS_ACTIVE
-            else:  # Pipeline not found
-                self.pipelines[pipeline_name] = False
-
-        self.logger.debug(f"Pipeline cache: {self.pipelines}")
-        return self.pipelines[pipeline_name]
-
     def retrieve_pipeline_execution(self, peh_id: str):
+        """Retrieve pipeline execution information and set as current execution in the Client
+
+        Arguments:
+            peh_id {str} -- Unique Pipeline execution ID
+
+        Returns:
+            None
+        """
         throw_none_or_empty(peh_id, "Pipeline is not specified")
 
         item = self.get_peh_record(peh_id)
@@ -293,4 +294,66 @@ class PipelineExecutionHistoryAPI:
         if not item["active"]:
             raise ValueError("Pipeline execution is inactive")
 
-        self.client.set_pipeline_execution(peh_id, item["pipeline"])
+        self.set_pipeline_execution(peh_id, item["pipeline"])
+
+    def end_pipeline_execution_failed(self, component: str = None, issue_comment: str = None) -> bool:
+        """Closes Pipeline Execution History record with FAILED status
+
+        Arguments:
+            component {str} -- Optional. Component of Pipeline Execution
+            issue_comment {str} -- Optional. Comment
+
+        Returns:
+            bool -- True if successful
+        """
+        return self.update_pipeline_execution(
+            PEH_STATUS_FAILED, component=component, issue_comment=issue_comment
+        )
+
+    def end_pipeline_execution_success(self, component: str = None) -> bool:
+        """Closes Pipeline Execution History record with COMPLETED status
+
+        Arguments:
+            component {str} -- Optional. Component of Pipeline Execution
+
+        Returns:
+            bool -- True if successful
+        """
+        return self.update_pipeline_execution(PEH_STATUS_COMPLETED, component=component)
+
+    def end_pipeline_execution_cancel(self, component: str = None, issue_comment: str = None) -> bool:
+        """Closes Pipeline execution with CANCELED status
+
+        Arguments:
+            component {str} -- Optional. Component of Pipeline Execution
+            issue_comment {str} -- Optional. Comment
+
+        Returns:
+            bool -- True if successful
+        """
+        return self.update_pipeline_execution(
+            PEH_STATUS_CANCELED, component=component, issue_comment=issue_comment
+        )
+
+    def is_pipeline_set(self) -> bool:
+        """Check if current pipeline execution is set
+
+        Returns:
+            bool -- True if pipeline execution is set
+        """
+        return self.pipeline_execution_id is not None and self.pipeline_name is not None
+
+    def reset_pipeline_execution(self):
+        """Clears the current pipeline execution"""
+        self.pipeline_execution_id = None
+        self.pipeline_name = None
+
+    def set_pipeline_execution(self, pipeline_execution_id: str, pipeline_name: str):
+        """Sets the current pipeline execution
+
+        Arguments:
+            pipeline_execution_id {str} -- Unique identifier of pipeline execution
+            pipeline_name {str} -- Pipeline name
+        """
+        self.pipeline_execution_id = pipeline_execution_id
+        self.pipeline_name = pipeline_name
